requirements.txt
pytrec-eval
cohere==4.36
voyageai
tqdm
vertexai
openai
transformers==4.51.0
tiktoken
scipy==1.10.1
gritlm
torchmetrics
pyserini
gensim